+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - You can now set attributes defining them in `data.attributes`, properties with `false` values will be removed
 - BREAKING CHANGE: `onCreate` and `onUpdate` methods have been converted to properties of the `options` object
   you can now use them with `new Popper(a, b, { ...otherOptions, onCreate: () => {}, onUpdate: () => {} })`
+- BREAKING CHANGE: `gpuAcceleration` is now an option of the `applyStyle` modifier (`modifiers.applyStyle.gpuAcceleration`)
 
 ## v0.6.4
 - 30a781e: Remove necessity of remove polyfill on IE (thanks @rafaelverger)